+++ extension/entrypoints/sidepanel/main.tsx
@@ -1,0 +1,5 @@
+import { render } from 'preact';
+import { App } from './App';
+import './styles/sidepanel.css';
+
+render(<App />, document.getElementById('app')!);